fix(jarparsecatalog): tmp here could have been null if it failed to created the tmp directory

+++ metadata/src/main/java/io/kaoto/backend/metadata/parser/JarParseCatalog.java
@@ -97,7 +97,9 @@ public class JarParseCatalog<T extends Metadata>
         }
 
         try {
-            FileUtils.deleteDirectory(tmp.toFile());
+            if (tmp != null) {
+                FileUtils.deleteDirectory(tmp.toFile());
+            }
             if (!location.equalsIgnoreCase(url)) {
                 FileUtils.deleteQuietly(new File(location));
             }
